+++ sw_projects/P3_app/InSpkrAudio.c
@@ -25,6 +25,8 @@
 #include <stdio.h>
 #include <string.h>
 #include <pthread.h>
+#include <sys/socket.h>
+#include <time.h>
 #include <syscall.h>
 #include "../common/saturnregisters.h"
 #include "../common/saturndrivers.h"
@@ -39,8 +41,13 @@
 #define VALIGNMENT 4096                             // buffer alignment
 #define VBASE 0x1000								// DMA start at 4K into buffer
 #define VDMATRANSFERSIZE 256                        // write 1 message at a time
-#define VMAXDMABATCHFRAMES 8                        // opportunistically batch queued UDP frames into one DMA
+#define VSPKNORMALQUEUEFRAMES 3                     // normal queue depth before a DMA write
+#define VSPKPREFILLQUEUEFRAMES 6                    // refill deeper after startup or underrun recovery
+#define VMAXDMABATCHFRAMES 8                        // max queued frames we coalesce into one DMA write
 #define VSTARTUPDELAY 100                           // 100 messages (~100ms) before reporting under or overflows
+#define VSPKPREFILLLOWFRAMES 4                      // re-enter prefill when FIFO occupancy falls below this
+#define VSPKPREFILLHIGHFRAMES 12                    // stay in prefill until FIFO occupancy reaches this
+#define VSPKMAXQUEUEAGEUS 2500U                     // bound extra speaker latency while still allowing coalescing
 
 static void NoteSpeakerUnderflow(bool ReportingEnabled, bool Underflowed, bool *UnderflowActive,
                                  unsigned int Current)
@@ -71,6 +78,29 @@ static void NoteSpeakerUnderflow(bool ReportingEnabled, bool Underflowed, bool *
         printf("Codec speaker FIFO Underflowed, depth now = %d\n", Current);
 }
 
+static uint64_t GetMonotonicTimeNs(void)
+{
+    struct timespec Now;
+
+    if (clock_gettime(CLOCK_MONOTONIC, &Now) != 0)
+        return 0;
+
+    return ((uint64_t)Now.tv_sec * 1000000000ULL) + (uint64_t)Now.tv_nsec;
+}
+
+static uint32_t GetSpeakerTargetFrames(unsigned int Current, bool *PrefillActive)
+{
+    const unsigned int LowWords = VSPKPREFILLLOWFRAMES * VMEMWORDSPERFRAME;
+    const unsigned int HighWords = VSPKPREFILLHIGHFRAMES * VMEMWORDSPERFRAME;
+
+    if (Current < LowWords)
+        *PrefillActive = true;
+    else if (Current >= HighWords)
+        *PrefillActive = false;
+
+    return *PrefillActive ? VSPKPREFILLQUEUEFRAMES : VSPKNORMALQUEUEFRAMES;
+}
+
 
 //
 // listener thread for incoming DDC (speaker) audio packets
@@ -81,11 +111,11 @@ static void NoteSpeakerUnderflow(bool ReportingEnabled, bool Underflowed, bool *
 void *IncomingSpkrAudio(void *arg)                      // listener thread
 {
     struct ThreadSocketData *ThreadData;                  // socket etc data for this thread
-    struct sockaddr_in addr_from;                         // holds MAC address of source of incoming messages
-    uint8_t UDPInBuffer[VSPEAKERAUDIOSIZE];               // incoming buffer
-    struct iovec iovecinst;                               // iovcnt buffer - 1 for each outgoing buffer
-    struct msghdr datagram;                               // multiple incoming message header
-    int size;                                             // UDP datagram length
+    uint8_t UDPInBuffers[VMAXDMABATCHFRAMES][VSPEAKERAUDIOSIZE];
+    struct iovec IovecList[VMAXDMABATCHFRAMES];
+    struct mmsghdr DatagramList[VMAXDMABATCHFRAMES];
+    unsigned int MsgIndex = 0;
+    int Received = 0;
 
 //
 // variables for DMA buffer 
@@ -96,18 +126,33 @@ void *IncomingSpkrAudio(void *arg)                      // listener thread
     uint32_t Depth = 0;
     int DMAWritefile_fd = -1;								// DMA read file device
     bool FIFOOverflow, FIFOUnderflow, FIFOOverThreshold;
-    uint32_t RegVal;
-    unsigned int Current;                                   // current occupied locations in FIFO
+    uint32_t RegVal = 0;
+    unsigned int Current = 0;                               // current occupied locations in FIFO
     unsigned int StartupCount = 0;                          // used to delay reporting of under & overflows
     bool PrevSDRActive = false;                             // used to detect change of state
     bool UnderflowActive = false;                           // tracks one continuous starvation episode
-    uint32_t BatchFrames = 0;
-    uint32_t BatchBytes = 0;
+    bool PrefillActive = true;                              // maintains a speaker FIFO cushion after startup or underrun
+    uint32_t PendingFrames = 0;
+    uint32_t PendingBytes = 0;
+    uint64_t PendingStartNs = 0;
+    uint32_t TargetFrames = VSPKPREFILLQUEUEFRAMES;
+    uint64_t QueueAgeUs = 0;
+    struct timespec ReceiveTimeout;
+    struct timespec *ReceiveTimeoutPtr = NULL;
 
 
     ThreadData = (struct ThreadSocketData *)arg;
     atomic_store(&ThreadData->Active, true);
     printf("spinning up speaker audio thread with port %u, pid=%ld\n", (unsigned int)atomic_load(&ThreadData->Portid), syscall(SYS_gettid));
+
+    memset(DatagramList, 0, sizeof(DatagramList));
+    for (MsgIndex = 0; MsgIndex < VMAXDMABATCHFRAMES; MsgIndex++)
+    {
+        IovecList[MsgIndex].iov_base = UDPInBuffers[MsgIndex];
+        IovecList[MsgIndex].iov_len = VSPEAKERAUDIOSIZE;
+        DatagramList[MsgIndex].msg_hdr.msg_iov = &IovecList[MsgIndex];
+        DatagramList[MsgIndex].msg_hdr.msg_iovlen = 1;
+    }
 
     //
     // setup DMA buffer
@@ -163,83 +208,129 @@ void *IncomingSpkrAudio(void *arg)                      // listener thread
         {
             StartupCount = VSTARTUPDELAY;
             UnderflowActive = false;
+            PrefillActive = true;
+            PendingFrames = 0;
+            PendingBytes = 0;
+            PendingStartNs = 0;
         }
         else if(!SDRActiveNow)
+        {
             UnderflowActive = false;
+            PrefillActive = true;
+            Current = 0;
+            PendingFrames = 0;
+            PendingBytes = 0;
+            PendingStartNs = 0;
+        }
         PrevSDRActive = SDRActiveNow;
 
-        memset(&iovecinst, 0, sizeof(struct iovec));            // clear buffers
-        memset(&datagram, 0, sizeof(datagram));
-        iovecinst.iov_base = &UDPInBuffer;                      // set buffer for incoming message number i
-        iovecinst.iov_len = VSPEAKERAUDIOSIZE;
-        datagram.msg_iov = &iovecinst;
-        datagram.msg_iovlen = 1;
-        datagram.msg_name = &addr_from;
-        datagram.msg_namelen = sizeof(addr_from);
-        //
-        // receive operation thread
-        //
-        size = recvmsg(atomic_load(&ThreadData->Socketid), &datagram, 0);   // get one message. If it times out, sets size=-1
-        if(size < 0 && errno != EAGAIN)
+        if(PendingFrames < VMAXDMABATCHFRAMES)
         {
-            perror("recvfrom fail, Speaker data");
-            P23PerfTelemetryCounterAdd(eP23PerfCounterSpkrRecvErrors, 1U);
-            atomic_store(&ThreadError, true);
-            break;
-        }
-        if(size == VSPEAKERAUDIOSIZE)                           // we have received a packet!
-        {
-            BatchFrames = 0;
-            BatchBytes = 0;
-            while((size == VSPEAKERAUDIOSIZE) && (BatchFrames < VMAXDMABATCHFRAMES))
+            unsigned int ReceiveGoal = VMAXDMABATCHFRAMES - PendingFrames;
+            int ReceiveFlags = MSG_WAITFORONE;
+
+            ReceiveTimeoutPtr = NULL;
+            if((PendingFrames != 0) && (PendingStartNs != 0))
             {
+                uint64_t RemainingAgeUs = 0;
+                uint64_t ElapsedAgeUs = (GetMonotonicTimeNs() - PendingStartNs) / 1000ULL;
+                if(ElapsedAgeUs < VSPKMAXQUEUEAGEUS)
+                    RemainingAgeUs = VSPKMAXQUEUEAGEUS - ElapsedAgeUs;
+                if(RemainingAgeUs == 0)
+                    Received = 0;
+                else
+                {
+                    ReceiveTimeout.tv_sec = (time_t)(RemainingAgeUs / 1000000ULL);
+                    ReceiveTimeout.tv_nsec = (long)((RemainingAgeUs % 1000000ULL) * 1000ULL);
+                    ReceiveTimeoutPtr = &ReceiveTimeout;
+                    Received = recvmmsg(atomic_load(&ThreadData->Socketid), DatagramList, ReceiveGoal, ReceiveFlags, ReceiveTimeoutPtr);
+                }
+            }
+            else
+                Received = recvmmsg(atomic_load(&ThreadData->Socketid), DatagramList, ReceiveGoal, ReceiveFlags, NULL);
+
+            if((Received < 0) && (errno != EAGAIN) && (errno != EWOULDBLOCK))
+            {
+                perror("recvfrom fail, Speaker data");
+                P23PerfTelemetryCounterAdd(eP23PerfCounterSpkrRecvErrors, 1U);
+                atomic_store(&ThreadError, true);
+                break;
+            }
+            if(Received < 0)
+                Received = 0;
+        }
+        else
+            Received = 0;
+
+        if(Received > 0)                                        // we have received one or more packets
+        {
+            for(MsgIndex = 0; MsgIndex < (unsigned int)Received; MsgIndex++)
+            {
+                if(DatagramList[MsgIndex].msg_len != VSPEAKERAUDIOSIZE)
+                    continue;
+                if(PendingFrames == 0)
+                    PendingStartNs = GetMonotonicTimeNs();
                 if(StartupCount != 0)                                   // decrement startup message count
                     StartupCount--;
                 atomic_store(&NewMessageReceived, true);
                 P23PerfTelemetryCounterAdd(eP23PerfCounterSpkrPackets, 1U);
                 P23PerfTelemetryCounterAdd(eP23PerfCounterSpkrBytes, VSPEAKERAUDIOSIZE);
                 RegVal += 1;            //debug
-                memcpy(SpkBasePtr + BatchBytes, UDPInBuffer + 4, VDMATRANSFERSIZE);              // copy out spk samples
-                BatchBytes += VDMATRANSFERSIZE;
-                BatchFrames++;
-                size = recvmsg(atomic_load(&ThreadData->Socketid), &datagram, MSG_DONTWAIT);
-                if(size < 0)
-                {
-                    if((errno == EAGAIN) || (errno == EWOULDBLOCK))
-                        break;
-                    perror("recvfrom fail while draining Speaker data");
-                    P23PerfTelemetryCounterAdd(eP23PerfCounterSpkrRecvErrors, 1U);
-                    atomic_store(&ThreadError, true);
+                memcpy(SpkBasePtr + PendingBytes, UDPInBuffers[MsgIndex] + 4, VDMATRANSFERSIZE); // queue speaker samples for the next DMA
+                PendingBytes += VDMATRANSFERSIZE;
+                PendingFrames++;
+                if(PendingFrames >= VMAXDMABATCHFRAMES)
                     break;
-                }
             }
-            if(atomic_load(&ThreadError))
-                break;
+        }
+
+        if(PendingFrames == 0)
+            continue;
+
+        QueueAgeUs = 0;
+        if(PendingStartNs != 0)
+            QueueAgeUs = (GetMonotonicTimeNs() - PendingStartNs) / 1000ULL;
+
+        TargetFrames = GetSpeakerTargetFrames(Current, &PrefillActive);
+        if((PendingFrames < TargetFrames) && (QueueAgeUs < VSPKMAXQUEUEAGEUS) && (PendingFrames < VMAXDMABATCHFRAMES))
+            continue;
+
+        {
             Depth = ReadFIFOMonitorChannel(eSpkCodecDMA, &FIFOOverflow, &FIFOOverThreshold, &FIFOUnderflow, &Current);        // read the FIFO free locations
             if((StartupCount == 0) && FIFOOverThreshold && UseDebug)
                 printf("Codec speaker FIFO Overthreshold, depth now = %d\n", Current);
+            if(FIFOUnderflow)
+                PrefillActive = true;
             NoteSpeakerUnderflow((StartupCount == 0) && SDRActiveNow, FIFOUnderflow, &UnderflowActive, Current);
 
-            while ((Depth < (VMEMWORDSPERFRAME * BatchFrames)) && !atomic_load(&ExitRequested))
+            while ((Depth < (VMEMWORDSPERFRAME * PendingFrames)) && !atomic_load(&ExitRequested))
             {
                 usleep(1000);
                 Depth = ReadFIFOMonitorChannel(eSpkCodecDMA, &FIFOOverflow, &FIFOOverThreshold, &FIFOUnderflow, &Current);
                 if((StartupCount == 0) && FIFOOverThreshold && UseDebug)
                     printf("Codec speaker FIFO Overthreshold, depth now = %d\n", Current);
+                if(FIFOUnderflow)
+                    PrefillActive = true;
                 NoteSpeakerUnderflow((StartupCount == 0) && SDRActiveNow, FIFOUnderflow, &UnderflowActive, Current);
             }
             if(atomic_load(&ExitRequested))
                 break;
-            if(BatchBytes != 0)
+            if(PendingBytes != 0)
             {
-                if(DMAWriteToFPGA(DMAWritefile_fd, SpkBasePtr, BatchBytes, VADDRSPKRSTREAMWRITE) < 0)
+                if(DMAWriteToFPGA(DMAWritefile_fd, SpkBasePtr, PendingBytes, VADDRSPKRSTREAMWRITE) < 0)
                 {
                     P23PerfTelemetryCounterAdd(eP23PerfCounterSpkrDMAErrors, 1U);
                     atomic_store(&ThreadError, true);
                     break;
                 }
                 P23PerfTelemetryCounterAdd(eP23PerfCounterSpkrDMAWrites, 1U);
-                P23PerfTelemetryCounterAdd(eP23PerfCounterSpkrDMAWriteBytes, BatchBytes);
+                P23PerfTelemetryCounterAdd(eP23PerfCounterSpkrDMAWriteBytes, PendingBytes);
+                Current += PendingFrames * VMEMWORDSPERFRAME;
+                if(Current >= (VSPKPREFILLHIGHFRAMES * VMEMWORDSPERFRAME))
+                    PrefillActive = false;
+                PendingFrames = 0;
+                PendingBytes = 0;
+                PendingStartNs = 0;
             }
         }
     }